+++ R/Core.R
@@ -45,13 +45,13 @@ NULL
 #' file <- system.file("extdata", "ExampleDb.gz", package="alakazam")
 #' 
 #' # Read all columns in and convert sequence fields to upper case
-#' df <- readChangeoDb(file)
+#' db <- readChangeoDb(file)
 #' 
 #' # Subset columns and convert sequence fields to upper case
-#' df <- readChangeoDb(file, select=c("SEQUENCE_ID", "SEQUENCE_IMGT"))
+#' db <- readChangeoDb(file, select=c("SEQUENCE_ID", "SEQUENCE_IMGT"))
 #' 
 #' # Drop columns and do not alter sequence field case
-#' df <- readChangeoDb(file, drop=c("D_CALL", "DUPCOUNT"), seq_upper=FALSE)
+#' db <- readChangeoDb(file, drop=c("D_CALL", "DUPCOUNT"), seq_upper=FALSE)
 #' 
 #' @export
 readChangeoDb <- function(file, select=NULL, drop=NULL, seq_upper=TRUE) {
@@ -60,7 +60,7 @@ readChangeoDb <- function(file, select=NULL, drop=NULL, seq_upper=TRUE) {
                      "GERMLINE_IMGT", "GERMLINE_IMGT_D_MASK")
     text_columns <- c("SEQUENCE_ID", "CLONE", "SAMPLE")
     
-    df <- tryCatch({ 
+    db <- tryCatch({ 
             # Attempt to use fread
             data.table::fread(file, sep="\t", header=TRUE, stringsAsFactors=FALSE,
                               na.strings=c("", "NA", "None"), select=select, drop=drop,
@@ -68,29 +68,29 @@ readChangeoDb <- function(file, select=NULL, drop=NULL, seq_upper=TRUE) {
         }, error=function(e) {
             # Read file using read.delim if fread fails
             #message("Failed to read file with fread. Falling back to read.table.")
-            df <- read.delim(file, as.is=TRUE, na.strings=c("", "NA", "None"))
+            db <- read.delim(file, as.is=TRUE, na.strings=c("", "NA", "None"))
             
             # Select columns
-            select_columns <- colnames(df)
+            select_columns <- colnames(db)
             if(!is.null(select)) { select_columns <- intersect(select_columns, select) }
             if(!is.null(drop)) { select_columns <- setdiff(select_columns, drop) }
             
-            subset(df, select=select_columns)
+            subset(db, select=select_columns)
         })    
 
     # Convert sequence fields to upper case
     if (seq_upper) {
-        for (x in intersect(seq_columns, names(df))) {
-            df[, x] <- toupper(df[, x]) 
+        for (x in intersect(seq_columns, names(db))) {
+            db[, x] <- toupper(db[, x]) 
         }
     }
     
     # Convert text fields to character
-    for (x in intersect(text_columns, names(df))) {
-        df[, x] <- as.character(df[, x])
+    for (x in intersect(text_columns, names(db))) {
+        db[, x] <- as.character(db[, x])
     }
     
-    return(df)
+    return(db)
 }